+++ Blockchain/Sources/Blockchain/Validator/ValidatorService.swift
@@ -7,13 +7,15 @@ public final class ValidatorService: Sendable {
     private let safrole: SafroleService
     private let extrinsicPool: ExtrinsicPoolService
     private let blockAuthor: BlockAuthor
+    private let dataAvailability: DataAvailability
 
     public init(
         blockchain: Blockchain,
         keystore: KeyStore,
         eventBus: EventBus,
         scheduler: Scheduler,
-        dataProvider: BlockchainDataProvider
+        dataProvider: BlockchainDataProvider,
+        dataStore: DataStore
     ) async {
         self.blockchain = blockchain
         self.keystore = keystore
@@ -37,6 +39,14 @@ public final class ValidatorService: Sendable {
             keystore: keystore,
             scheduler: scheduler,
             extrinsicPool: extrinsicPool
+        )
+
+        dataAvailability = await DataAvailability(
+            config: blockchain.config,
+            eventBus: eventBus,
+            scheduler: scheduler,
+            dataProvider: dataProvider,
+            dataStore: dataStore
         )
     }
