+++ Node/Sources/Node/Node.swift
@@ -8,63 +8,7 @@ import Utils
 
 private let logger = Logger(label: "node")
 
-public typealias RPCConfig = Server.Config
-public typealias NetworkConfig = Network.Config
-
-public enum Database {
-    case inMemory
-    case rocksDB(path: URL)
-
-    public func open(chainspec: ChainSpec) async throws -> BlockchainDataProvider {
-        switch self {
-        case let .rocksDB(path):
-            logger.info("Using RocksDB backend at \(path.absoluteString)")
-            let backend = try await RocksDBBackend(
-                path: path,
-                config: chainspec.getConfig(),
-                genesisBlock: chainspec.getBlock(),
-                genesisStateData: chainspec.getState()
-            )
-            return try await BlockchainDataProvider(backend)
-        case .inMemory:
-            logger.info("Using in-memory backend")
-            let genesisBlock = try chainspec.getBlock()
-            let genesisStateData = try chainspec.getState()
-            let backend = try StateBackend(InMemoryBackend(), config: chainspec.getConfig(), rootHash: Data32())
-            try await backend.writeRaw(Array(genesisStateData))
-            let genesisState = try await State(backend: backend)
-            let genesisStateRef = StateRef(genesisState)
-            return try await BlockchainDataProvider(InMemoryDataProvider(genesisState: genesisStateRef, genesisBlock: genesisBlock))
-        }
-    }
-}
-
 public class Node {
-    public struct Config {
-        public var rpc: RPCConfig?
-        public var network: NetworkConfig
-        public var peers: [NetAddr]
-        public var local: Bool
-        public var name: String?
-        public var database: Database
-
-        public init(
-            rpc: RPCConfig?,
-            network: NetworkConfig,
-            peers: [NetAddr] = [],
-            local: Bool = false,
-            name: String? = nil,
-            database: Database = .inMemory
-        ) {
-            self.rpc = rpc
-            self.network = network
-            self.peers = peers
-            self.local = local
-            self.name = name
-            self.database = database
-        }
-    }
-
     public let config: Config
     public let blockchain: Blockchain
     public let rpcServer: Server?
